finance_news_bot/requirements.txt
requests
beautifulsoup4
schedule
python-dotenv
feedparser
